Suggestion for restructuring on component file-structure
