+++ src/lib/components/HistoryMatch/index.js
@@ -1,0 +1,1 @@
+export { default } from './HistoryMatch';
