+++ src/lib/components/HistoryMatch/utils/legend.js
