+++ src/lib/components/LayeredMap/components/CanvasOverlay.react.js
