+++ src/lib/components/LayeredMap/components/Colormap.react.js
