+++ src/lib/components/LayeredMap/components/CompositeMapLayer.react.js
