+++ src/lib/components/LayeredMap/components/DrawControls.react.js
