+++ src/lib/components/LayeredMap/components/Switch.react.js
