+++ src/lib/components/LayeredMap/components/ValueInfoBox.react.js
