+++ src/lib/components/LayeredMap/components/VerticalZoom.react.js
