+++ src/lib/components/LayeredMap/shaders/fragmentShaderWithHillshading.fs.glsl
