+++ src/lib/components/LayeredMap/shaders/fragmentShaderWithoutHillshading.fs.glsl
