+++ src/lib/components/Map/utils/cell.js
