+++ src/lib/components/Map/utils/color-scale.js
