+++ src/lib/components/Map/utils/compass.js
