+++ src/lib/components/Map/utils/component.js
