+++ src/lib/components/Map/utils/distance-scale.js
