+++ src/lib/components/Map/utils/flow_animation.js
