+++ src/lib/components/Map/utils/grid.js
