+++ src/lib/components/Map/utils/linear_algebra.js
