+++ src/lib/components/Map/utils/map.js
