+++ src/lib/components/Map/utils/particle_generator.js
