+++ src/lib/components/Map/utils/tests/arbitraries.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-magic-numbers */
-import jsc from "jsverify";
+import jsc from "./jsverify";
 import { affineIndependent, Vector, Matrix } from "../linear_algebra";
 import { Cell } from "../cell";
 import Grid from "../grid";
