+++ src/lib/components/Map/utils/tests/cell.spec.js
@@ -1,4 +1,4 @@
-import jsc from "jsverify";
+import jsc from "./jsverify";
 import {
     arbitraryCell,
     arbitraryNonEmptyCell,
