+++ src/lib/components/Map/utils/vertical-slider.js
