+++ src/lib/shared/component.js
