+++ src/lib/shared/slider.css
