+++ src/lib/shared/slider.js
